fix security issues and add telegram phone login
Fixes IDOR on cart/order/address/notifications, admin panel auth,
JWT access/refresh token confusion, login/register enumeration,
broken route paths, money fields (Float->BigInteger), and several
crash bugs (branches update, product update, celery retry).

Adds passwordless phone login via Telegram bot (webhook + OTP),
rate limiting on auth endpoints, and fixes Dockerfile to install
exact locked dependency versions.

+++ app/schemas/order.py
@@ -19,7 +19,7 @@ class OrderUpdateRequest(BaseModel):
     address_id: int
     promocode_id: int
     branch_id: int
-    total_price: float
+    total_price: int
 
 
 class OrderListResponse(BaseModel):
@@ -28,7 +28,7 @@ class OrderListResponse(BaseModel):
     address_id: int
     promocode_id: int | None = None
     branch_id: int
-    total_price: float
+    total_price: int
     created_at: datetime
 
     model_config = {
@@ -48,14 +48,11 @@ class OrderListResponse(BaseModel):
     }
 
 
-from pydantic import BaseModel
-
-
 class OrderItemResponse(BaseModel):
     id: int
     product_id: int
     quantity: int
-    price: float
+    price: int
 
 
 class OrderCreateResponse(BaseModel):
@@ -64,7 +61,7 @@ class OrderCreateResponse(BaseModel):
     address_id: int
     branch_id: int
     promocode_id: int | None
-    total_price: float
+    total_price: int
     order_items: list[OrderItemResponse]
